+++ cmd/policy/policy.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/soluble-ai/soluble-cli/pkg/api"
 	"github.com/soluble-ai/soluble-cli/pkg/log"
+	"github.com/soluble-ai/soluble-cli/pkg/policy/custompolicybuilder"
 	"github.com/soluble-ai/soluble-cli/pkg/policy/manager"
 	"github.com/soluble-ai/soluble-cli/pkg/tools"
 	"github.com/soluble-ai/soluble-cli/pkg/util"
@@ -31,38 +32,21 @@ func Command() *cobra.Command {
 }
 
 func createCommand() *cobra.Command {
-	m := &manager.M{}
-	pt := &manager.PolicyTemplate{}
+	cpb := &custompolicybuilder.PolicyTemplate{}
 	c := &cobra.Command{
 		Use:   "create",
 		Short: "Create custom policy. Generates skeleton policy and metadata file",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if err := pt.ValidateCreateInput(); err != nil {
+			if err := cpb.ValidateCreateInput(); err != nil {
 				return err
 			}
-			if err := pt.CreateCustomPolicyTemplate(); err != nil {
+			if err := cpb.CreateCustomPolicyTemplate(); err != nil {
 				return err
 			}
 			return nil
 		},
 	}
-	m.RunOpts.Register(c)
-	flags := c.Flags()
-	flags.StringVar(&pt.PolicyName, "name", "", "name of policy to create")
-	_ = c.MarkFlagRequired("name")
-	flags.StringVar(&pt.CheckType, "check-type", "", "policy target")
-	_ = c.MarkFlagRequired("check-type")
-	flags.StringVar(&pt.PolicyType, "type", "", "policy type")
-	_ = c.MarkFlagRequired("type")
-
-	// Optional
-	flags.StringVarP(&pt.PolicyDir, "directory", "d", "policies", "path to custom policies directory")
-	flags.StringVar(&pt.PolicyDesc, "description", "", "policy description")
-	flags.StringVar(&pt.PolicyTitle, "title", "", "policy title")
-	flags.StringVar(&pt.PolicySeverity, "severity", "medium", "policy severity")
-	flags.StringVar(&pt.PolicyCategory, "category", "", "policy category")
-	flags.StringVar(&pt.PolicyRsrcType, "resource-type", "", "policy resource type")
-
+	cpb.Register(c)
 	return c
 }
