+++ pkg/policy/custompolicybuilder/custom-policy-builder.go
@@ -1,4 +1,4 @@
-package manager
+package custompolicybuilder
 
 import (
 	"fmt"
